Add loader utils file
